API Updates
Now the website will only fetch 5 posts at a time, then if the user scrolls down to the bottom of the page it will request more to be loaded, this should prevent too much lag on the website.

On the edit post feature on the dashboard it can now render the CURRENT image that is currently on the post.

Started moving the API to a google VM, I have the google VM completely setup I just need to finish transferring everything to the VM