Announcement Page Updated
Added a share button to all posts on the announcements page and if you click the share button it will copy a link to your clipboard and the link will take you to an entirely separate page with just that post, it should be good for like kids who want to share their post with their parents.

Added PostHog analytics so I can see what areas of the site people use the most

Also fixed a bunch of bugs and added the announcements page the the navbar so people will actually find it.